+++ src/dfx/src/commands/canister/delete.rs
@@ -7,10 +7,14 @@ use crate::lib::operations::canister;
 use crate::lib::operations::canister::{
     deposit_cycles, start_canister, stop_canister, update_settings,
 };
+use crate::lib::operations::cycles_ledger::{
+    wallet_deposit_to_cycles_ledger, CYCLES_LEDGER_ENABLED,
+};
 use crate::lib::root_key::fetch_root_key_if_needed;
 use crate::util::assets::wallet_wasm;
 use crate::util::blob_from_arguments;
-use anyhow::Context;
+use crate::util::clap::parsers::icrc_subaccount_parser;
+use anyhow::{bail, Context};
 use candid::Principal;
 use clap::Parser;
 use dfx_core::canister::build_wallet_canister;
@@ -23,6 +27,7 @@ use ic_utils::interfaces::management_canister::builders::InstallMode;
 use ic_utils::interfaces::management_canister::CanisterStatus;
 use ic_utils::interfaces::ManagementCanister;
 use ic_utils::Argument;
+use icrc_ledger_types::icrc1::account::{Account, Subaccount};
 use num_traits::cast::ToPrimitive;
 use slog::info;
 use std::convert::TryFrom;
@@ -71,6 +76,11 @@ pub struct CanisterDeleteOpts {
     /// Auto-confirm deletion for a non-stopped canister.
     #[arg(long, short)]
     yes: bool,
+
+    /// Subaccount of the selected identity to deposit cycles to.
+    //TODO(SDK-1331): unhide
+    #[arg(long, value_parser = icrc_subaccount_parser, hide = true)]
+    to_subaccount: Option<Subaccount>,
 }
 
 #[context("Failed to delete canister '{}'.", canister)]
@@ -83,6 +93,7 @@ async fn delete_canister(
     withdraw_cycles_to_canister: Option<String>,
     withdraw_cycles_to_dank: bool,
     withdraw_cycles_to_dank_principal: Option<String>,
+    to_cycles_ledger_subaccount: Option<Subaccount>,
 ) -> DfxResult {
     let log = env.get_logger();
     let mut canister_id_store = env.get_canister_id_store()?;
@@ -99,19 +110,23 @@ async fn delete_canister(
         let to_dank = withdraw_cycles_to_dank || withdraw_cycles_to_dank_principal.is_some();
 
         // Get the canister to transfer the cycles to.
-        let target_canister_id = if no_withdrawal {
-            None
+        let withdraw_target = if no_withdrawal {
+            WithdrawTarget::NoWithdrawal
         } else if to_dank {
-            Some(DANK_PRINCIPAL)
+            WithdrawTarget::Dank
         } else {
             match withdraw_cycles_to_canister {
                 Some(ref target_canister_id) => {
-                    Some(Principal::from_text(target_canister_id).with_context(|| {
-                        format!("Failed to read canister id {:?}.", target_canister_id)
-                    })?)
+                    let canister_id =
+                        Principal::from_text(target_canister_id).with_context(|| {
+                            format!("Failed to read canister id {:?}.", target_canister_id)
+                        })?;
+                    WithdrawTarget::Canister { canister_id }
                 }
                 None => match call_sender {
-                    CallSender::Wallet(wallet_id) => Some(*wallet_id),
+                    CallSender::Wallet(wallet_id) => WithdrawTarget::Canister {
+                        canister_id: *wallet_id,
+                    },
                     CallSender::SelectedId => {
                         let network = env.get_network_descriptor();
                         let agent_env = create_agent_environment(env, Some(network.name.clone()))?;
@@ -120,7 +135,19 @@ async fn delete_canister(
                             .expect("No selected identity.")
                             .to_string();
                         // If there is no wallet, then do not attempt to withdraw the cycles.
-                        wallet_canister_id(network, &identity_name)?
+                        match wallet_canister_id(network, &identity_name)? {
+                            Some(canister_id) => WithdrawTarget::Canister { canister_id },
+                            None if CYCLES_LEDGER_ENABLED => {
+                                let Some(my_principal)  =  env.get_selected_identity_principal() else { bail!("Identity has no principal attached") };
+                                WithdrawTarget::CyclesLedger {
+                                    to: Account {
+                                        owner: my_principal,
+                                        subaccount: to_cycles_ledger_subaccount,
+                                    },
+                                }
+                            }
+                            _ => WithdrawTarget::NoWithdrawal,
+                        }
                     }
                 },
             }
@@ -135,11 +162,10 @@ async fn delete_canister(
         };
         fetch_root_key_if_needed(env).await?;
 
-        if let Some(target_canister_id) = target_canister_id {
+        if withdraw_target != WithdrawTarget::NoWithdrawal {
             info!(
                 log,
-                "Beginning withdrawal of cycles to canister {}; on failure try --no-wallet --no-withdrawal.",
-                target_canister_id
+                "Beginning withdrawal of cycles; on failure try --no-wallet --no-withdrawal."
             );
 
             // Determine how many cycles we can withdraw.
@@ -197,40 +223,55 @@ async fn delete_canister(
                         break;
                     }
                     let cycles_to_withdraw = cycles - margin;
-                    let result = if !to_dank {
-                        info!(
-                            log,
-                            "Attempting to transfer {} cycles to canister {}.",
-                            cycles_to_withdraw,
-                            target_canister_id
-                        );
-                        // Transfer cycles from the source canister to the target canister using the temporary wallet.
-                        deposit_cycles(
-                            env,
-                            target_canister_id,
-                            &CallSender::Wallet(canister_id),
-                            cycles_to_withdraw,
-                        )
-                        .await
-                    } else {
-                        info!(
-                            log,
-                            "Attempting to transfer {} cycles to dank principal {}.",
-                            cycles_to_withdraw,
-                            dank_target_principal
-                        );
-                        let wallet = build_wallet_canister(canister_id, agent).await?;
-                        let opt_principal = Some(dank_target_principal);
-                        wallet
-                            .call(
+                    let result = match withdraw_target {
+                        WithdrawTarget::NoWithdrawal => Ok(()),
+                        WithdrawTarget::Dank => {
+                            info!(
+                                log,
+                                "Attempting to transfer {} cycles to dank principal {}.",
+                                cycles_to_withdraw,
+                                dank_target_principal
+                            );
+                            let wallet = build_wallet_canister(canister_id, agent).await?;
+                            let opt_principal = Some(dank_target_principal);
+                            wallet
+                                .call(
+                                    DANK_PRINCIPAL,
+                                    "mint",
+                                    Argument::from_candid((opt_principal,)),
+                                    cycles_to_withdraw,
+                                )
+                                .call_and_wait()
+                                .await
+                                .context("Failed mint call.")
+                        }
+                        WithdrawTarget::Canister {
+                            canister_id: target_canister_id,
+                        } => {
+                            info!(
+                                log,
+                                "Attempting to transfer {} cycles to canister {}.",
+                                cycles_to_withdraw,
+                                target_canister_id
+                            );
+                            // Transfer cycles from the source canister to the target canister using the temporary wallet.
+                            deposit_cycles(
+                                env,
                                 target_canister_id,
-                                "mint",
-                                Argument::from_candid((opt_principal,)),
+                                &CallSender::Wallet(canister_id),
                                 cycles_to_withdraw,
                             )
-                            .call_and_wait()
                             .await
-                            .context("Failed mint call.")
+                        }
+                        WithdrawTarget::CyclesLedger { to } => {
+                            wallet_deposit_to_cycles_ledger(
+                                agent,
+                                canister_id,
+                                cycles_to_withdraw,
+                                to,
+                            )
+                            .await
+                        }
                     };
                     if result.is_ok() {
                         info!(log, "Successfully withdrew {} cycles.", cycles_to_withdraw);
@@ -239,7 +280,7 @@ async fn delete_canister(
                         info!(log, "Not enough margin. Trying again with more margin.");
                         attempts += 1;
                     } else {
-                        // Unforseen error. Report it back to user
+                        // Unforeseen error. Report it back to user
                         result?;
                     }
                 }
@@ -293,6 +334,7 @@ pub async fn exec(
             opts.withdraw_cycles_to_canister,
             opts.withdraw_cycles_to_dank,
             opts.withdraw_cycles_to_dank_principal,
+            opts.to_subaccount,
         )
         .await
     } else if opts.all {
@@ -307,6 +349,7 @@ pub async fn exec(
                     opts.withdraw_cycles_to_canister.clone(),
                     opts.withdraw_cycles_to_dank,
                     opts.withdraw_cycles_to_dank_principal.clone(),
+                    opts.to_subaccount,
                 )
                 .await?;
             }
@@ -315,4 +358,12 @@ pub async fn exec(
     } else {
         unreachable!()
     }
+}
+
+#[derive(Debug, Clone, Copy, Eq, PartialEq)]
+enum WithdrawTarget {
+    NoWithdrawal,
+    Dank,
+    CyclesLedger { to: Account },
+    Canister { canister_id: Principal },
 }